Adding DemoBehaviorChooser and changing function signatures to use double instead of float for time in the Behavior system

+++ basestation/include/anki/cozmo/basestation/behaviors/behaviorLookForFaces.h
@@ -30,11 +30,11 @@ namespace Cozmo {
     virtual Result Init() override;
     
     // Always runnable for now?
-    virtual bool IsRunnable(float currentTime_sec) const override { return true; }
+    virtual bool IsRunnable(double currentTime_sec) const override { return true; }
     
-    virtual Status Update(float currentTime_sec) override;
+    virtual Status Update(double currentTime_sec) override;
     
-    virtual Result Interrupt(float currentTime_sec) override;
+    virtual Result Interrupt(double currentTime_sec) override;
     
     virtual const std::string& GetName() const override {
       static const std::string name("LookForFaces");
